+++ src/factory/FilterFactory.js
@@ -12,18 +12,19 @@ Util.ext(_FilterFactory.prototype,{
 		var rs = new this.types[type].clazz(this.baseClass,component);
 		Util.extProp(rs,this.types[type].props);
 
-		this.instances.push(rs);
-
-		//inject
-		var services = null;
-		if(this.types[type].services){
-			services = [];
-			for(var i=0;i<this.types[type].services.length;i++){
-				var serv = ServiceFactory.newInstanceOf(this.types[type].services[i],rs);
-				services.push(serv);
+		if(rs.onCreate){
+			//inject
+			var services = null;
+			if(this.types[type].services){
+				services = [];
+				for(var i=0;i<this.types[type].services.length;i++){
+					var serv = ServiceFactory.newInstanceOf(this.types[type].services[i],rs);
+					services.push(serv);
+				}
 			}
+			
+			services ? rs.onCreate.apply(rs,services) : rs.onCreate();
 		}
-		rs.onCreate && rs.onCreate.apply(rs,services);
 
 		return rs;
 	}