+++ src/factory/ServiceFactory.js
@@ -18,21 +18,24 @@ Util.ext(_ServiceFactory.prototype,{
 			rs = this.types[type].singleton;
 		}else{
 			rs = new this.types[type].clazz(this.baseClass);
-			this.instances.push(rs);
 			Util.extProp(rs,this.types[type].props);
 		}
 
-		//inject
-		var services = null;
-		if(this.types[type].services){
-			services = [];
-			for(var i=0;i<this.types[type].services.length;i++){
-				var serv = ServiceFactory.newInstanceOf(this.types[type].services[i],rs);
-				services.push(serv);
-			}
-		}
 		rs.$host = host;
-		rs.onCreate && rs.onCreate.apply(rs,services);
+
+		if(rs.onCreate){
+			//inject
+			var services = null;
+			if(this.types[type].services){
+				services = [];
+				for(var i=0;i<this.types[type].services.length;i++){
+					var serv = ServiceFactory.newInstanceOf(this.types[type].services[i],rs);
+					services.push(serv);
+				}
+			}
+			
+			services ? rs.onCreate.apply(rs,services) : rs.onCreate();
+		}		
 
 		return rs;
 	}